DefaultStrategy for Sequence has been implmented
