+++ moderncollections/src/main/java/com/rwalker/HowToFunction.java
@@ -1,0 +1,6 @@
+package com.rwalker;
+
+public enum HowToFunction {
+    QUEUE,
+    STACK
+}
