+++ moderncollections/src/test/java/com/rwalker/StackTest.java
@@ -51,7 +51,7 @@ public class StackTest {
         Sequence<Integer> test = new Sequence<Integer>();
         test.push(10);
         test.push(20);
-        assertEquals(20, (int) test.peek(Sequence.HowToFunction.STACK));
+        assertEquals(20, (int) test.peek(HowToFunction.STACK));
     }
 
     /**
@@ -60,7 +60,7 @@ public class StackTest {
     @Test(expected = NullPointerException.class)
     public void testPeekEmpty() {
         Sequence<Integer> test = new Sequence<Integer>();
-        test.peek(Sequence.HowToFunction.STACK);
+        test.peek(HowToFunction.STACK);
     }
 
     /**
